+++ webui/react/src/components/ProjectMoveModal.module.scss
@@ -1,11 +1,3 @@
-.base {
-  display: flex;
-  flex-direction: column;
-
-  .label {
-    color: var(--theme-float-on);
-  }
-}
 .workspaceOptionDisabled {
   display: flex;
   justify-content: space-between;
