--- webui/react/src/hooks/useModal/Project/useModalProjectCreate.tsx
@@ -1,110 +1,0 @@
-import { ModalFuncProps } from 'antd/es/modal/Modal';
-import React, { useCallback, useEffect, useMemo } from 'react';
-
-import Form from 'components/kit/Form';
-import Input from 'components/kit/Input';
-import { paths } from 'routes/utils';
-import { createProject } from 'services/api';
-import useModal, { ModalHooks } from 'shared/hooks/useModal/useModal';
-import { DetError, ErrorLevel, ErrorType } from 'shared/utils/error';
-import { routeToReactUrl } from 'shared/utils/routes';
-import handleError from 'utils/error';
-
-const FORM_ID = 'new-project-form';
-
-interface FormInputs {
-  description?: string;
-  projectName: string;
-}
-
-interface Props {
-  onClose?: () => void;
-  workspaceId: number;
-}
-
-const useModalProjectCreate = ({ onClose, workspaceId }: Props): ModalHooks => {
-  const [form] = Form.useForm<FormInputs>();
-  const { modalOpen: openOrUpdate, modalRef, ...modalHook } = useModal({ onClose });
-  const projectName = Form.useWatch('projectName', form);
-
-  const modalContent = useMemo(() => {
-    return (
-      <Form autoComplete="off" form={form} id={FORM_ID} layout="vertical">
-        <Form.Item
-          label="Project Name"
-          name="projectName"
-          rules={[{ message: 'Name is required', required: true }]}>
-          <Input maxLength={80} />
-        </Form.Item>
-        <Form.Item label="Description" name="description">
-          <Input />
-        </Form.Item>
-      </Form>
-    );
-  }, [form]);
-
-  const handleOk = useCallback(async () => {
-    const values = await form.validateFields();
-
-    try {
-      if (values) {
-        const response = await createProject({
-          description: values.description,
-          name: values.projectName,
-          workspaceId,
-        });
-        routeToReactUrl(paths.projectDetails(response.id));
-        form.resetFields();
-      }
-    } catch (e) {
-      if (e instanceof DetError) {
-        handleError(e, {
-          level: e.level,
-          publicMessage: e.publicMessage,
-          publicSubject: 'Unable to create project.',
-          silent: false,
-          type: e.type,
-        });
-      } else {
-        handleError(e, {
-          level: ErrorLevel.Error,
-          publicMessage: 'Please try again later.',
-          publicSubject: 'Unable to create project.',
-          silent: false,
-          type: ErrorType.Server,
-        });
-      }
-    }
-  }, [form, workspaceId]);
-
-  const getModalProps = useMemo((): ModalFuncProps => {
-    return {
-      closable: true,
-      content: modalContent,
-      icon: null,
-      okButtonProps: { disabled: !projectName, form: FORM_ID, htmlType: 'submit' },
-      okText: 'Create Project',
-      onOk: handleOk,
-      title: 'New Project',
-    };
-  }, [handleOk, modalContent, projectName]);
-
-  const modalOpen = useCallback(
-    (initialModalProps: ModalFuncProps = {}) => {
-      openOrUpdate({ ...getModalProps, ...initialModalProps });
-    },
-    [getModalProps, openOrUpdate],
-  );
-
-  /**
-   * When modal props changes are detected, such as modal content
-   * title, and buttons, update the modal.
-   */
-  useEffect(() => {
-    if (modalRef.current) openOrUpdate(getModalProps);
-  }, [getModalProps, modalRef, openOrUpdate]);
-
-  return { modalOpen, modalRef, ...modalHook };
-};
-
-export default useModalProjectCreate;
