--- webui/react/src/hooks/useModal/Project/useModalProjectEdit.module.scss
@@ -1,10 +1,0 @@
-.base {
-  display: flex;
-  flex-direction: column;
-  gap: 16px;
-
-  .label {
-    color: var(--theme-float-on);
-    margin-bottom: -8px;
-  }
-}
