--- webui/react/src/hooks/useModal/Project/useModalProjectNoteDelete.module.scss
@@ -1,9 +1,0 @@
-.base {
-  display: flex;
-  flex-direction: column;
-  gap: 16px;
-
-  :global(p) {
-    margin: 0;
-  }
-}
